+++ src/Graphics/VulkanContextFactory.cpp
@@ -1,12 +1,8 @@
 #include "VulkanContextFactory.h"
 
 #include "VulkanContext.h"
-#include "VulkanWrappers/Memory/Buffer.h"
-#include "VulkanWrappers/Memory/VmaAllocator.h"
+#include "VulkanContextFactoryFunctions.h"
 
-#include <GLFW/glfw3.h>
-
-#include <VkBootstrap.h>
 #include <Vulkan/vulkan.hpp>
 #include <fmt/format.h>
 
@@ -19,139 +15,6 @@ VULKAN_HPP_DEFAULT_DISPATCH_LOADER_DYNAMIC_STORAGE
 namespace trg
 {
 	extern std::atomic_bool alreadyCreated = false;
-
-	namespace VulkanContextFactoryPrivate
-	{
-		const char* toString(VkDebugUtilsMessageTypeFlagsEXT messageType)
-		{
-			switch(messageType)
-			{
-				case(7):
-					return "General | Validation | Performance";
-				case(6):
-					return "Validation | Performance";
-				case(5):
-					return "General | Performance";
-				case(4):
-					return "Performance";
-				case(3):
-					return "General | Validation";
-				case(2):
-					return "Validation";
-				case(1):
-					return "General";
-				default:
-					return "Unknown";
-			}
-		}
-
-		vk::Bool32 debugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
-								 VkDebugUtilsMessageTypeFlagsEXT messageType,
-								 const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
-								 void*)
-		{
-			auto messageSeverityString = vk::to_string(static_cast<vk::DebugUtilsMessageSeverityFlagBitsEXT>(messageSeverity));
-
-			auto message = fmt::format("[{0} : {1}]\n{2}\n", messageSeverityString, toString(messageType), pCallbackData->pMessage);
-
-			std::cout << message << std::endl;
-
-			return false;
-		}
-
-		vkb::Instance makeInstance()
-		{
-			auto instanceResult = vkb::InstanceBuilder()
-									  .require_api_version(1, 2)
-									  .enable_extension(VK_EXT_DEBUG_UTILS_EXTENSION_NAME)
-									  .request_validation_layers()
-									  .build();
-
-			if(!instanceResult.has_value())
-			{
-				throw std::runtime_error(fmt::format("Instance initialization failed. Error code: {0}. Error message: {1}",
-													 instanceResult.error().value(),
-													 instanceResult.error().message()));
-			}
-
-			return instanceResult.value();
-		}
-
-		auto makeDebugMessenger(vk::Instance& instance)
-		{
-			auto messageTypes = vk::DebugUtilsMessageTypeFlagBitsEXT::eGeneral | vk::DebugUtilsMessageTypeFlagBitsEXT::ePerformance |
-								vk::DebugUtilsMessageTypeFlagBitsEXT::eValidation;
-			auto severity = vk::DebugUtilsMessageSeverityFlagBitsEXT::eWarning | vk::DebugUtilsMessageSeverityFlagBitsEXT::eError;
-
-			auto debugMessengerCreateInfo =
-				vk::DebugUtilsMessengerCreateInfoEXT(vk::DebugUtilsMessengerCreateFlagBitsEXT(0), severity, messageTypes, &debugCallback);
-
-			return instance.createDebugUtilsMessengerEXT(debugMessengerCreateInfo);
-		}
-
-		vkb::PhysicalDevice getPhysicalDevice(const vkb::Instance& instance, GLFWwindow& window)
-		{
-			VkSurfaceKHR surface;
-			VkResult result = glfwCreateWindowSurface(instance.instance, &window, nullptr, &surface);
-
-			auto physicalDeviceResult = vkb::PhysicalDeviceSelector(instance).set_surface(surface).select();
-
-			if(!physicalDeviceResult.has_value())
-			{
-				throw std::runtime_error(fmt::format("Physical device initialization failed. Error code: {0}. Error message: {1}",
-													 physicalDeviceResult.error().value(),
-													 physicalDeviceResult.error().message()));
-			}
-
-			return physicalDeviceResult.value();
-		}
-
-		vkb::Device makeDevice(const vkb::PhysicalDevice& physicalDevice)
-		{
-			auto features = vk::PhysicalDeviceVulkan12Features();
-			features.separateDepthStencilLayouts = true;
-
-			auto deviceResult = vkb::DeviceBuilder(physicalDevice).add_pNext(&features).build();
-
-			if(!deviceResult.has_value())
-			{
-				throw std::runtime_error(fmt::format("Device initialization failed. Error code: {0}. Error message: {1}",
-													 deviceResult.error().value(),
-													 deviceResult.error().message()));
-			}
-
-			return deviceResult.value();
-		}
-
-		vkb::Swapchain makeSwapchain(const vkb::Device& device)
-		{
-			auto swapChainResult = vkb::SwapchainBuilder(device).build();
-
-			if(!swapChainResult.has_value())
-			{
-				throw std::runtime_error(fmt::format("Swapchain initialization failed. Error code: {0}. Error message: {1}",
-													 swapChainResult.error().value(),
-													 swapChainResult.error().message()));
-			}
-
-			return swapChainResult.value();
-		}
-
-		CommandQueue makeGraphicsQueue(const vkb::Device& device)
-		{
-			auto queueResult = device.get_queue(vkb::QueueType::graphics);
-
-			if(!queueResult.has_value())
-			{
-				throw std::runtime_error(fmt::format("Graphics queue initialization failed. Error code: {0}. Error message: {1}",
-													 queueResult.error().value(),
-													 queueResult.error().message()));
-			}
-
-			return CommandQueue(vk::Queue(queueResult.value()), device.get_queue_index(vkb::QueueType::graphics).value());
-		}
-	}  // namespace VulkanContextFactoryPrivate
-
 
 	VulkanContextFactory::VulkanContextFactory()
 	{
@@ -172,9 +35,8 @@ namespace trg
 			context->m_swapchainExtent = vk::Extent2D(1024, 768);
 			context->m_window =
 				glfwCreateWindow(context->m_swapchainExtent.width, context->m_swapchainExtent.height, "Throne", nullptr, nullptr);
-
-			using namespace VulkanContextFactoryPrivate;
-
+			glfwSetWindowUserPointer(context->m_window, context.get());
+			glfwSetFramebufferSizeCallback(context->m_window, &glfwFrameBufferResizeCallback);
 
 			auto instance = makeInstance();
 			context->m_instance = vk::Instance(instance.instance);
@@ -186,17 +48,16 @@ namespace trg
 			context->m_physicalDevice = vk::PhysicalDevice(physicalDevice.physical_device);
 			context->m_surface = vk::SurfaceKHR(physicalDevice.surface);
 
-			auto device = makeDevice(physicalDevice);
-			context->m_device = vk::Device(device.device);
+			context->m_vkbDevice = std::make_unique<vkb::Device>(makeDevice(physicalDevice));
+			context->m_device = vk::Device(context->m_vkbDevice->device);
 			vk::defaultDispatchLoaderDynamic.init(context->m_device);
 
 			initializeVmaDefaultAllocator(context->m_physicalDevice, context->m_device);
 
-			auto swapchain = makeSwapchain(device);
+			auto swapchain = makeSwapchain(*context->m_vkbDevice);
 			new(&context->m_swapchain) Swapchain(context->m_device, swapchain);
-			context->m_swapchains = {context->m_swapchain.getSwapchain()};
 
-			context->m_graphicsQueue = makeGraphicsQueue(device);
+			context->m_graphicsQueue = makeGraphicsQueue(*context->m_vkbDevice);
 
 			return context;
 		}
